Complete 2/3 TestProject

+++ Assets/_Project/Logic/Gameplay/ConfigsScripts/EnemyConfig.cs
@@ -5,9 +5,9 @@ namespace _Project.Logic.Gameplay.ConfigsScripts
     [CreateAssetMenu(fileName = "EnemyConfig", menuName = "Gameplay/EnemyConfig")]
     public class EnemyConfig: ScriptableObject
     {
-        [field: SerializeField] public float MaxHealth { get; private set; }
+        [field: SerializeField] public int MaxHealth { get; private set; }
         [field: SerializeField] public float Speed { get; private set; }
-        [field: SerializeField] public float Damage { get; private set; }
+        [field: SerializeField] public int Damage { get; private set; }
         [field:SerializeField] public int Reward { get; private set; }
     }
 }
